jd02-02 Task B completed

+++ src/by/it/ragach/jd02_02/entity/Pensioner.java
@@ -6,7 +6,7 @@ public class Pensioner extends Customer{
 
     public final String name;
 
-    public Pensioner(int number) {
+    public Pensioner(long number) {
         super(number);
         this.name = "Customer (Pensioner) №" + number;
     }
